Fix Resume Inbox sort buttons: root-caused to 4 real bugs + a backfill
- 'Sort by Newest Added' was a pure no-op against the API's own
  already-newest-first default order. Rewritten as a genuine
  ascending/descending toggle that actually reverses the list.
- match_requisition()'s skill-based fallback compared whole skill
  phrases against single title words via exact set intersection
  ('sap fico' & 'sap' is always empty) - the real reason almost nothing
  in the queue had ever been auto-scored at intake. Fixed to check
  substring containment instead.
- That same query had no is_active/status filter, so a tenant with
  heavy test-suite activity could have its entire top-50-by-created_at
  window filled with soft-deleted rows, crowding out every real open
  requisition regardless of skill match quality.
- Backfilling the existing queue with the fixed matcher surfaced a
  genuinely new, previously-latent crash: total_years_exp (NUMERIC ->
  Python Decimal via asyncpg) multiplied against a float in the
  composite readiness formula. Fixed with a single float() cast.

Backfilled 15 of 46 previously-unmatched real candidates with real
applications + scores using the exact same functions live intake uses.

+++ backend/routers/ner.py
@@ -213,7 +213,16 @@ def score_candidate(
     skill_score = round(min(skill_similarity * 100, 100), 2)
 
     # 2. Experience fit
-    actual_yr = (parsed.get('total_years_exp') or 0) or (candidate_exp_mo / 12)
+    # Real bug fixed 2026-08-20, found while backfilling requisition
+    # matching for the existing Resume Inbox queue: total_years_exp comes
+    # from a NUMERIC column (candidate_parsed_data), which asyncpg returns
+    # as a Python Decimal - a genuinely pre-existing latent bug (the old
+    # composite formula multiplied exp_score by a float too), just never
+    # triggered before today because no candidate with a Decimal-valued
+    # total_years_exp had ever been scored via this exact code path.
+    # float() here keeps every downstream computation in plain float,
+    # since Decimal arithmetic with float literals raises TypeError.
+    actual_yr = float((parsed.get('total_years_exp') or 0) or (candidate_exp_mo / 12))
     if required_exp_yr_max:
         if actual_yr < required_exp_yr_min:
             exp_score = max(0, 50 - (required_exp_yr_min - actual_yr) * 10)
